pyproject.toml
[tool.poetry]
name = "blagues_api"
version = "1.0.2"
description = "Official client for Blagues API"
homepage = "https://www.blagues-api.fr/"
authors = [
  "baptiste0928 <contact@baptiste0928.net>",
  "DraftMan <contact@draftman.fr>"
]
license = "MIT"
readme = "README.md"
repository = "https://github.com/Blagues-API/package-py"

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "^3.7"
pydantic = "^1.8"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
pytest-asyncio = "^0.15.1"
python-dotenv = "^0.17.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode="auto"
